SHaNEWEB/requirements.txt
Flask==3.1.2
mysql-connector-python==8.3.0